Change the path variable to a literal path string
Had to change the path from a dynamic value to a literal path, may try to change again later.

It did run, not EVERYTHING in the downloads folder was moved, I think it had an issue when it ran into folders, need to make sure it can move those too.